# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample fiber project server.",
        "title": "Machine task",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/admin/login": {
            "post": {
                "description": "Authenticate an admin and return a token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Admin login",
                "parameters": [
                    {
                        "description": "Admin login data",
                        "name": "adminLogin",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestmodels.AdminLogin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully authenticated",
                        "schema": {
                            "$ref": "#/definitions/responsemodels.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/responsemodels.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responsemodels.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "requestmodels.AdminLogin": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 5
                }
            }
        },
...
        "responsemodels.Response": {
            "type": "object",
            "properties": {
                "error": {},
                "message": {
                    "type": "string"
                },
                "result": {},
                "status_code": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "authorization": {
            "description": "IMPORTANT: TYPE \"Bearer\" FOLLOWED BY A SPACE AND JWT TOKEN.",
            "type": "apiKey",
            "name": "authorization",
            "in": "header"
        }
    }
}